[FIX] #273 edit orderbook chart ui
