+++ src/components/trade/CurrentPrice.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="flex items-baseline gap-x-2">
     <BaseTypography size="2xl" weight="bold" class="mb-3">
-      {{ currentPrice.toLocaleString() }}원
+      {{ (currentPrice ?? 0).toLocaleString() }}원
     </BaseTypography>
     <!-- 추후 거래차트 시 데이터 넣기 -->
     <BaseTypography>(2.39%)</BaseTypography>
@@ -9,34 +9,12 @@
 </template>
 
 <script setup>
-import { ref, watch } from 'vue'
 import BaseTypography from '../common/Typography/BaseTypography.vue'
-import { getOrderBookByFundingId } from '@/api/orderbook'
 
-const props = defineProps({
-  fundingId: {
-    type: [String, Number],
-    required: true,
+defineProps({
+  currentPrice: {
+    type: Number,
+    default: 0,
   },
 })
-
-const currentPrice = ref(0)
-
-const fetchCurrentPrice = async (id) => {
-  try {
-    const res = await getOrderBookByFundingId(Number(id))
-    currentPrice.value = res.data?.data?.currentPrice || 0
-  } catch (err) {
-    console.error('📉 현재가 조회 실패:', err)
-  }
-}
-
-// ✅ fundingId 변경 시 즉시 호출
-watch(
-  () => props.fundingId,
-  (newId) => {
-    fetchCurrentPrice(newId)
-  },
-  { immediate: true },
-)
 </script>
